+++ .changeset/short-moose-wait.md
@@ -1,0 +1,5 @@
+---
+'@seedcord/services': minor
+---
+
+better Logger with new utils and better file transports + lots of config options
